+++ App/Navigators/LoginNavigator.js
@@ -1,17 +1,16 @@
 import React from 'react';
-import { View,  Text } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import {Provider} from 'react-redux';
-import { configStore } from '../reducers/store';
+import {View, Text} from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';
+import {createStackNavigator} from '@react-navigation/stack';
+
 import LoginClient from '../Containers/LoginClient';
-import { RegisterScreen } from '../Containers/RegisterClient';
+import {RegisterScreen} from '../Containers/RegisterClient';
 
 const Stack = createStackNavigator();
 
 function SelectRole() {
   return (
-    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+    <View style={{flex: 1, alignItems: 'center', justifyContent: 'center'}}>
       <Text>Role Screen</Text>
     </View>
   );
@@ -19,15 +18,13 @@ function SelectRole() {
 
 const LoginNavigator = () => {
   return (
-    <Provider store={configStore()}>
-			<NavigationContainer>
-				<Stack.Navigator>
-					<Stack.Screen name="Login" component={LoginClient} />
-					<Stack.Screen name="Register" component={RegisterScreen} />
-					<Stack.Screen name="SelecRole" component={SelectRole} />
-				</Stack.Navigator>
-			</NavigationContainer>
-    </Provider>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="Login" component={LoginClient} />
+        <Stack.Screen name="Register" component={RegisterScreen} />
+        <Stack.Screen name="SelecRole" component={SelectRole} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 };
