+++ android/app/src/main/java/com/nextline/MainApplication.java
@@ -27,7 +27,7 @@ public class MainApplication extends Application implements ReactApplication {
           List<ReactPackage> packages = new PackageList(this).getPackages();
           // Packages that cannot be autolinked yet can be added manually here, for example:
           // packages.add(new MyReactNativePackage());
-          packages.add(new RNGestureHandlerPackage());
+          // packages.add(new RNGestureHandlerPackage());
           return packages;
         }
